requirements.txt
accelerate==0.33.0
fastapi==0.113.0
Flask==3.0.3
groq==0.11.0
huggingface-hub==0.24.5
Jinja2==3.1.4
keras==3.4.1
matplotlib==3.9.2
matplotlib-inline==0.1.7
numpy==1.26.4
opencv-python==4.9.0.80
opencv-python-headless==4.10.0.84
pandas==2.2.2
peft==0.13.2
pillow==10.4.0
py3-tts==3.5
pydantic==2.9.0
pydantic_core==2.23.2
pyttsx3==2.91
requests==2.32.3
safetensors==0.4.3
scipy==1.14.1
seaborn==0.13.2
streamlit==1.38.0
tensorflow==2.17.0
tokenizers==0.19.1
torch==2.4.0
torchaudio==2.4.0
torchvision==0.19.0
tqdm==4.66.5
transformers==4.43.3
ultralytics==8.2.82
uvicorn==0.30.6
watchfiles==0.24.0
wcwidth==0.2.13
websockets==13.1
yolov5==7.0.13